+++ nix/scufris.nix
@@ -36,6 +36,9 @@
   };
   service = native.service;
   ctl = native.ctl;
+  staging = import ./staging.nix {
+    inherit pkgs self service desktop;
+  };
   devShell = import ./dev-shell.nix {inherit pkgs voice;};
   docs = import ./docs.nix {inherit inputs self pkgs;};
 in {
@@ -48,6 +51,7 @@ in {
     desktop
     service
     ctl
+    staging
     devShell
     docs
     ;